trading_bot/trading_bot/requirements.txt
python-binance
colorama
tabulate
numpy
python-dotenv
